ref: Split sources.rs from types.rs
This creates a new sources.rs file to hold types and impls of them
related to download sources.  It also moves the FileId in the objects
actor to sources::SourceFileId so this is all together.
