+++ src/actors/objects/filesystem.rs
@@ -5,21 +5,22 @@ use std::sync::Arc;
 use futures01::{Future, IntoFuture};
 
 use crate::actors::objects::common::prepare_download_paths;
-use crate::actors::objects::{DownloadPath, DownloadStream, FileId, ObjectError, ObjectErrorKind};
-use crate::types::{FileType, FilesystemSourceConfig, ObjectId};
+use crate::actors::objects::{DownloadStream, ObjectError, ObjectErrorKind};
+use crate::sources::{FileType, FilesystemSourceConfig, SourceFileId, SourceLocation};
+use crate::types::ObjectId;
 
 pub(super) fn prepare_downloads(
     source: &Arc<FilesystemSourceConfig>,
     filetypes: &'static [FileType],
     object_id: &ObjectId,
-) -> Box<dyn Future<Item = Vec<FileId>, Error = ObjectError>> {
+) -> Box<dyn Future<Item = Vec<SourceFileId>, Error = ObjectError>> {
     let ids = prepare_download_paths(
         object_id,
         filetypes,
         &source.files.filters,
         source.files.layout,
     )
-    .map(|download_path| FileId::Filesystem(source.clone(), download_path))
+    .map(|download_path| SourceFileId::Filesystem(source.clone(), download_path))
     .collect();
 
     Box::new(Ok(ids).into_future())
@@ -27,7 +28,7 @@ pub(super) fn prepare_downloads(
 
 pub(super) fn download_from_source(
     source: Arc<FilesystemSourceConfig>,
-    download_path: &DownloadPath,
+    download_path: &SourceLocation,
 ) -> Box<dyn Future<Item = Option<DownloadStream>, Error = ObjectError>> {
     let download_abspath = source.path.join(&download_path.0);
     log::debug!("Fetching debug file from {:?}", download_abspath);
