+++ src/actors/objects/s3.rs
@@ -8,8 +8,9 @@ use rusoto_s3::S3;
 use tokio::codec::{BytesCodec, FramedRead};
 
 use crate::actors::objects::common::prepare_download_paths;
-use crate::actors::objects::{DownloadPath, DownloadStream, FileId, ObjectError, ObjectErrorKind};
-use crate::types::{FileType, ObjectId, S3SourceConfig, S3SourceKey};
+use crate::actors::objects::{DownloadStream, ObjectError, ObjectErrorKind};
+use crate::sources::{FileType, S3SourceConfig, S3SourceKey, SourceFileId, SourceLocation};
+use crate::types::ObjectId;
 
 lazy_static::lazy_static! {
     static ref AWS_HTTP_CLIENT: rusoto_core::HttpClient = rusoto_core::HttpClient::new().unwrap();
@@ -53,21 +54,21 @@ pub(super) fn prepare_downloads(
     source: &Arc<S3SourceConfig>,
     filetypes: &'static [FileType],
     object_id: &ObjectId,
-) -> Box<dyn Future<Item = Vec<FileId>, Error = ObjectError>> {
+) -> Box<dyn Future<Item = Vec<SourceFileId>, Error = ObjectError>> {
     let ids = prepare_download_paths(
         object_id,
         filetypes,
         &source.files.filters,
         source.files.layout,
     )
-    .map(|download_path| FileId::S3(source.clone(), download_path))
+    .map(|download_path| SourceFileId::S3(source.clone(), download_path))
     .collect();
     Box::new(Ok(ids).into_future())
 }
 
 pub(super) fn download_from_source(
     source: Arc<S3SourceConfig>,
-    download_path: &DownloadPath,
+    download_path: &SourceLocation,
 ) -> Box<dyn Future<Item = Option<DownloadStream>, Error = ObjectError>> {
     let key = {
         let prefix = source.prefix.trim_matches(&['/'][..]);
