+++ src/endpoints/applecrashreport.rs
@@ -11,7 +11,8 @@ use sentry_actix::ActixWebHubExt;
 use crate::actors::symbolication::SymbolicationActor;
 use crate::app::{ServiceApp, ServiceState};
 use crate::endpoints::symbolicate::SymbolicationRequestQueryParams;
-use crate::types::{RequestId, Scope, SourceConfig, SymbolicationResponse};
+use crate::sources::SourceConfig;
+use crate::types::{RequestId, Scope, SymbolicationResponse};
 use crate::utils::futures::ThreadPool;
 use crate::utils::multipart::{read_multipart_file, read_multipart_sources};
 use crate::utils::sentry::{SentryFutureExt, WriteSentryScope};
